+++ engine/engine_types.go
@@ -64,5 +64,5 @@ type Settings struct {
 
 	// Dispatch system settings
 	EnableDispatcher        bool
-	DispatchMaxWorkerAmount int64
+	DispatchMaxWorkerAmount int
 }
